updated to production 3a2580b3

+++ new/js/commonfunc.js
@@ -40,7 +40,10 @@ function clearStorage(keepOrg)
     var loadOrgKey = localStorage.loadOrgKey || "";
     var loadPhonegap = localStorage.isPhonegap || "";
     var userKey = localStorage.userKey || "";
-    localStorage.clear();
+    if (!window.dontClearCache) localStorage.clear();
+    localStorage.userKey = "";
+    localStorage.userOrgKey = "";
+    localStorage.userInstanceKey = "";
     //localStorage.removeItem('userOrgKey');
     //localStorage.removeItem('userOrg');
     //localStorage.removeItem('userInstanceKey');
@@ -60,6 +63,7 @@ function clearStorage(keepOrg)
 
 function initOrgPreferences(value)
 {
+    return;
     var prefs = plugins.appPreferences;
     var suitePrefs = prefs.iosSuite("group.io.sherpadesk.mobile");
     suitePrefs.store (ok, fail, 'org', value);
